Revert "Merge main into "Verification Refactoring - Part #2" (#377)"
This reverts commit b80fceebbd0b9978c33eb6e19520f68b2ed77bcb.
