+++ packages/mobile/src/app/verificationMigration.test.ts
@@ -5,7 +5,7 @@ import { verificationMigrationRan } from 'src/app/actions'
 import { numberVerifiedSelector, ranVerificationMigrationSelector } from 'src/app/selectors'
 import { runVerificationMigration } from 'src/app/verificationMigration'
 import { e164NumberToSaltSelector } from 'src/identity/reducer'
-import { e164NumberSelector, komenciContextSelector } from 'src/verify/module'
+import { e164NumberSelector, komenciContextSelector } from 'src/verify/reducer'
 import { getContractKitAsync } from 'src/web3/contracts'
 import { mtwAddressSelector } from 'src/web3/selectors'
 import { mockAccount, mockE164Number, mockE164NumberPepper } from 'test/values'
