+++ packages/mobile/src/identity/actions.ts
@@ -1,5 +1,6 @@
 import { normalizeAddressWith0x } from '@celo/base'
 import { E164Number } from '@celo/utils/lib/io'
+import { CodeInputStatus } from 'src/components/CodeInput'
 import {
   AddressToDisplayNameType,
   AddressToE164NumberType,
@@ -8,13 +9,25 @@ import {
   E164NumberToSaltType,
   WalletToAccountAddressType,
 } from 'src/identity/reducer'
-import { ContactMatches, ImportContactsStatus } from 'src/identity/types'
+import { ContactMatches, ImportContactsStatus, VerificationStatus } from 'src/identity/types'
+import { AttestationCode, CodeInputType } from 'src/identity/verification'
 import { Recipient } from 'src/recipients/recipient'
 
 export enum Actions {
+  START_VERIFICATION = 'IDENTITY/START_VERIFICATION',
+  CANCEL_VERIFICATION = 'IDENTITY/CANCEL_VERIFICATION',
+  RESET_VERIFICATION = 'IDENTITY/RESET_VERIFICATION',
+  SET_VERIFICATION_STATUS = 'IDENTITY/SET_VERIFICATION_STATUS',
+  SET_SEEN_VERIFICATION_NUX = 'IDENTITY/SET_SEEN_VERIFICATION_NUX',
+  SET_COMPLETED_CODES = 'IDENTITY/SET_COMPLETED_CODES',
+  REVOKE_VERIFICATION = 'IDENTITY/REVOKE_VERIFICATION',
+  REVOKE_VERIFICATION_STATE = 'IDENTITY/REVOKE_VERIFICATION_STATE',
+  FEELESS_REVOKE_VERIFICATION_STATE = 'IDENTITY/FEELESS_REVOKE_VERIFICATION_STATE',
+  RECEIVE_ATTESTATION_MESSAGE = 'IDENTITY/RECEIVE_ATTESTATION_MESSAGE',
+  INPUT_ATTESTATION_CODE = 'IDENTITY/INPUT_ATTESTATION_CODE',
+  COMPLETE_ATTESTATION_CODE = 'IDENTITY/COMPLETE_ATTESTATION_CODE',
   UPDATE_E164_PHONE_NUMBER_ADDRESSES = 'IDENTITY/UPDATE_E164_PHONE_NUMBER_ADDRESSES',
-  UPDATE_WALLET_TO_ACCOUNT_ADDRESS = 'IDENTITY/UPDATE_WALLET_TO_ACCOUNT_ADDRESS',
-  REMOVE_WALLET_TO_ACCOUNT_ADDRESS = 'IDENTITY/REMOVE_WALLET_TO_ACCOUNT_ADDRESS',
+  UPDATE_WALLET_TO_ACCOUNT_ADDRESS = 'UPDATE_WALLET_TO_ACCOUNT_ADDRESS',
   UPDATE_E164_PHONE_NUMBER_SALT = 'IDENTITY/UPDATE_E164_PHONE_NUMBER_SALT',
   UPDATE_KNOWN_ADDRESSES = 'IDENTITY/UPDATE_KNOWN_ADDRESSES',
   FETCH_ADDRESSES_AND_VALIDATION_STATUS = 'IDENTITY/FETCH_ADDRESSES_AND_VALIDATION_STATUS',
@@ -31,6 +44,66 @@ export enum Actions {
   REQUIRE_SECURE_SEND = 'IDENTITY/REQUIRE_SECURE_SEND',
   FETCH_DATA_ENCRYPTION_KEY = 'IDENTITY/FETCH_DATA_ENCRYPTION_KEY',
   UPDATE_ADDRESS_DEK_MAP = 'IDENTITY/UPDATE_ADDRESS_DEK_MAP',
+  RESEND_ATTESTATIONS = 'IDENTITY/RESEND_ATTESTATIONS',
+  SET_LAST_REVEAL_ATTEMPT = 'IDENTITY/SET_LAST_REVEAL_ATTEMPT',
+  REPORT_REVEAL_STATUS = 'IDENTITY/REPORT_REVEAL_STATUS',
+  SET_ATTESTATION_INPUT_STATUS = 'IDENTITY/SET_ATTESTATION_INPUT_STATUS',
+}
+
+export interface StartVerificationAction {
+  type: Actions.START_VERIFICATION
+  withoutRevealing: boolean | undefined
+  e164Number: string
+}
+
+export interface SetVerificationStatusAction {
+  type: Actions.SET_VERIFICATION_STATUS
+  status: VerificationStatus
+}
+
+export interface SetHasSeenVerificationNux {
+  type: Actions.SET_SEEN_VERIFICATION_NUX
+  status: boolean
+}
+
+export interface CancelVerificationAction {
+  type: Actions.CANCEL_VERIFICATION
+}
+
+export interface ResetVerificationAction {
+  type: Actions.RESET_VERIFICATION
+}
+
+export interface RevokeVerificationAction {
+  type: Actions.REVOKE_VERIFICATION
+}
+
+export interface RevokeVerificationStateAction {
+  type: Actions.REVOKE_VERIFICATION_STATE
+  walletAddress: string
+}
+
+export interface ReceiveAttestationMessageAction {
+  type: Actions.RECEIVE_ATTESTATION_MESSAGE
+  message: string
+  inputType: CodeInputType
+  index?: number
+}
+
+export interface SetCompletedCodesAction {
+  type: Actions.SET_COMPLETED_CODES
+  numComplete: number
+}
+
+export interface InputAttestationCodeAction {
+  type: Actions.INPUT_ATTESTATION_CODE
+  code: AttestationCode
+  index?: number
+}
+
+export interface CompleteAttestationCodeAction {
+  type: Actions.COMPLETE_ATTESTATION_CODE
+  code: AttestationCode
 }
 
 export interface UpdateE164PhoneNumberAddressesAction {
@@ -42,11 +115,6 @@ export interface UpdateE164PhoneNumberAddressesAction {
 export interface UpdateWalletToAccountAddressAction {
   type: Actions.UPDATE_WALLET_TO_ACCOUNT_ADDRESS
   walletToAccountAddress: WalletToAccountAddressType
-}
-
-export interface RemoveWalletToAccountAddressAction {
-  type: Actions.REMOVE_WALLET_TO_ACCOUNT_ADDRESS
-  walletAddress: string
 }
 
 export interface UpdateE164PhoneNumberSaltAction {
@@ -137,15 +205,42 @@ export interface UpdateAddressDekMapAction {
   dataEncryptionKey: string | null
 }
 
-// export interface SetLastRevealAttempt {
-// type: Actions.SET_LAST_REVEAL_ATTEMPT
-// time: number
-// }
+export interface ResendAttestations {
+  type: Actions.RESEND_ATTESTATIONS
+}
+
+export interface SetLastRevealAttempt {
+  type: Actions.SET_LAST_REVEAL_ATTEMPT
+  time: number
+}
+
+export interface ReportRevealStatusAction {
+  type: Actions.REPORT_REVEAL_STATUS
+  attestationServiceUrl: string
+  account: string
+  issuer: string
+  e164Number: string
+  pepper: string
+}
+
+export interface SetAttestationStatusAction {
+  type: Actions.SET_ATTESTATION_INPUT_STATUS
+  index: number
+  status: CodeInputStatus
+}
 
 export type ActionTypes =
+  | StartVerificationAction
+  | CancelVerificationAction
+  | ResetVerificationAction
+  | SetVerificationStatusAction
+  | SetHasSeenVerificationNux
+  | SetCompletedCodesAction
+  | ReceiveAttestationMessageAction
+  | InputAttestationCodeAction
+  | CompleteAttestationCodeAction
   | UpdateE164PhoneNumberAddressesAction
   | UpdateWalletToAccountAddressAction
-  | RemoveWalletToAccountAddressAction
   | UpdateE164PhoneNumberSaltAction
   | UpdateKnownAddressesAction
   | ImportContactsAction
@@ -161,6 +256,78 @@ export type ActionTypes =
   | EndFetchingAddressesAction
   | FetchDataEncryptionKeyAction
   | UpdateAddressDekMapAction
+  | ResendAttestations
+  | SetLastRevealAttempt
+  | ReportRevealStatusAction
+  | RevokeVerificationStateAction
+  | SetAttestationStatusAction
+
+export const startVerification = (
+  e164Number: string,
+  withoutRevealing: boolean | undefined
+): StartVerificationAction => ({
+  type: Actions.START_VERIFICATION,
+  withoutRevealing,
+  e164Number,
+})
+
+export const cancelVerification = (): CancelVerificationAction => ({
+  type: Actions.CANCEL_VERIFICATION,
+})
+
+export const resetVerification = (): ResetVerificationAction => ({
+  type: Actions.RESET_VERIFICATION,
+})
+
+export const setVerificationStatus = (status: VerificationStatus): SetVerificationStatusAction => ({
+  type: Actions.SET_VERIFICATION_STATUS,
+  status,
+})
+
+export const setHasSeenVerificationNux = (status: boolean): SetHasSeenVerificationNux => ({
+  type: Actions.SET_SEEN_VERIFICATION_NUX,
+  status,
+})
+
+export const revokeVerification = (): RevokeVerificationAction => ({
+  type: Actions.REVOKE_VERIFICATION,
+})
+
+// Will properly clear verification state when called
+export const revokeVerificationState = (walletAddress: string): RevokeVerificationStateAction => ({
+  type: Actions.REVOKE_VERIFICATION_STATE,
+  walletAddress,
+})
+
+export const receiveAttestationMessage = (
+  message: string,
+  inputType: CodeInputType,
+  index?: number
+): ReceiveAttestationMessageAction => ({
+  type: Actions.RECEIVE_ATTESTATION_MESSAGE,
+  message,
+  inputType,
+  index,
+})
+
+export const setCompletedCodes = (numComplete: number): SetCompletedCodesAction => ({
+  type: Actions.SET_COMPLETED_CODES,
+  numComplete,
+})
+
+export const inputAttestationCode = (
+  code: AttestationCode,
+  index?: number
+): InputAttestationCodeAction => ({
+  type: Actions.INPUT_ATTESTATION_CODE,
+  code,
+  index,
+})
+
+export const completeAttestationCode = (code: AttestationCode): CompleteAttestationCodeAction => ({
+  type: Actions.COMPLETE_ATTESTATION_CODE,
+  code,
+})
 
 export const fetchAddressesAndValidate = (
   e164Number: string,
@@ -206,13 +373,6 @@ export const updateWalletToAccountAddress = (
     walletToAccountAddress: newWalletToAccountAddresses,
   }
 }
-
-export const removeWalletToAccountAddress = (
-  walletAddress: string
-): RemoveWalletToAccountAddressAction => ({
-  type: Actions.REMOVE_WALLET_TO_ACCOUNT_ADDRESS,
-  walletAddress,
-})
 
 export const updateE164PhoneNumberSalts = (
   e164NumberToSalt: E164NumberToSaltType
@@ -312,4 +472,39 @@ export const updateAddressDekMap = (
   type: Actions.UPDATE_ADDRESS_DEK_MAP,
   address,
   dataEncryptionKey,
+})
+
+export const resendAttestations = (): ResendAttestations => ({
+  type: Actions.RESEND_ATTESTATIONS,
+})
+
+export const setLastRevealAttempt = (time: number): SetLastRevealAttempt => ({
+  type: Actions.SET_LAST_REVEAL_ATTEMPT,
+  time,
+})
+
+export const reportRevealStatus = (
+  attestationServiceUrl: string,
+  account: string,
+  issuer: string,
+  e164Number: string,
+  pepper: string
+): ReportRevealStatusAction => {
+  return {
+    type: Actions.REPORT_REVEAL_STATUS,
+    attestationServiceUrl,
+    account,
+    issuer,
+    e164Number,
+    pepper,
+  }
+}
+
+export const setAttestationInputStatus = (
+  index: number,
+  status: CodeInputStatus
+): SetAttestationStatusAction => ({
+  type: Actions.SET_ATTESTATION_INPUT_STATUS,
+  index,
+  status,
 })
