+++ packages/mobile/src/identity/reducer.ts
@@ -3,12 +3,19 @@ import dotProp from 'dot-prop-immutable'
 import { RehydrateAction } from 'redux-persist'
 import { createSelector } from 'reselect'
 import { Actions as AccountActions, ClearStoredAccountAction } from 'src/account/actions'
+import { CodeInputStatus } from 'src/components/CodeInput'
 import { Actions, ActionTypes } from 'src/identity/actions'
-import { ContactMatches, ImportContactsStatus } from 'src/identity/types'
+import { ContactMatches, ImportContactsStatus, VerificationStatus } from 'src/identity/types'
 import { removeKeyFromMapping } from 'src/identity/utils'
+import { AttestationCode, NUM_ATTESTATIONS_REQUIRED } from 'src/identity/verification'
 import { getRehydratePayload, REHYDRATE } from 'src/redux/persist-helper'
 import { RootState } from 'src/redux/reducers'
 import { StoreLatestInRecentsAction } from 'src/send/actions'
+import Logger from 'src/utils/Logger'
+import { isCodeRepeated } from 'src/verify/utils'
+
+export const ATTESTATION_CODE_PLACEHOLDER = 'ATTESTATION_CODE_PLACEHOLDER'
+export const ATTESTATION_ISSUER_PLACEHOLDER = 'ATTESTATION_ISSUER_PLACEHOLDER'
 
 export interface AddressToE164NumberType {
   [address: string]: string | null
@@ -72,7 +79,16 @@ export interface SecureSendDetails {
 }
 
 export interface State {
+  attestationCodes: AttestationCode[]
+  // we store acceptedAttestationCodes to tell user if code
+  // was already used even after Actions.RESET_VERIFICATION
+  acceptedAttestationCodes: AttestationCode[]
+  // Represents the status in the UI. Should be of size 3.
+  attestationInputStatus: CodeInputStatus[]
   // numCompleteAttestations is controlled locally
+  numCompleteAttestations: number
+  verificationStatus: VerificationStatus
+  hasSeenVerificationNux: boolean
   addressToE164Number: AddressToE164NumberType
   // Note: Do not access values in this directly, use the `getAddressFromPhoneNumber` helper in contactMapping
   e164NumberToAddress: E164NumberToAddressType
@@ -90,9 +106,20 @@ export interface State {
   // Contacts found during the matchmaking process
   matchedContacts: ContactMatches
   secureSendPhoneNumberMapping: SecureSendPhoneNumberMapping
+  lastRevealAttempt: number | null
 }
 
 const initialState: State = {
+  attestationCodes: [],
+  acceptedAttestationCodes: [],
+  attestationInputStatus: [
+    CodeInputStatus.Inputting,
+    CodeInputStatus.Disabled,
+    CodeInputStatus.Disabled,
+  ],
+  numCompleteAttestations: 0,
+  verificationStatus: VerificationStatus.Stopped,
+  hasSeenVerificationNux: false,
   addressToE164Number: {},
   e164NumberToAddress: {},
   walletToAccountAddress: {},
@@ -107,6 +134,7 @@ const initialState: State = {
   },
   matchedContacts: {},
   secureSendPhoneNumberMapping: {},
+  lastRevealAttempt: null,
 }
 
 export const reducer = (
@@ -121,13 +149,75 @@ export const reducer = (
       return {
         ...state,
         ...rehydratedState,
+        verificationStatus: VerificationStatus.Stopped,
         importContactsProgress: {
           status: ImportContactsStatus.Stopped,
           current: 0,
           total: 0,
         },
+        attestationInputStatus: initialState.attestationInputStatus,
       }
     }
+    case Actions.RESET_VERIFICATION:
+      return {
+        ...state,
+        attestationCodes: [],
+        numCompleteAttestations: 0,
+        verificationStatus: VerificationStatus.Stopped,
+      }
+    case Actions.REVOKE_VERIFICATION_STATE:
+      return {
+        ...state,
+        attestationCodes: [],
+        acceptedAttestationCodes: [],
+        numCompleteAttestations: 0,
+        verificationStatus: VerificationStatus.Stopped,
+        lastRevealAttempt: null,
+        walletToAccountAddress: removeKeyFromMapping(
+          state.walletToAccountAddress,
+          action.walletAddress
+        ),
+      }
+    case Actions.SET_VERIFICATION_STATUS:
+      return {
+        ...state,
+        verificationStatus: action.status,
+      }
+    case Actions.SET_SEEN_VERIFICATION_NUX:
+      return {
+        ...state,
+        hasSeenVerificationNux: action.status,
+      }
+    case Actions.SET_COMPLETED_CODES:
+      return {
+        ...state,
+        ...completeCodeReducer(state, action.numComplete),
+      }
+    case Actions.INPUT_ATTESTATION_CODE:
+      const codeAlreadyAdded = state.attestationCodes.some((code) => code.code === action.code.code)
+      const attestationCodes = codeAlreadyAdded
+        ? state.attestationCodes
+        : [...state.attestationCodes, action.code]
+      const attestationInputStatus = action.index
+        ? updatedInputStatuses(
+            state,
+            action.index,
+            codeAlreadyAdded || attestationCodes[action.index]?.code !== action.code.code
+              ? CodeInputStatus.Error
+              : CodeInputStatus.Processing
+          )
+        : state.attestationInputStatus
+      return {
+        ...state,
+        attestationCodes,
+        attestationInputStatus,
+      }
+    case Actions.COMPLETE_ATTESTATION_CODE:
+      return {
+        ...state,
+        numCompleteAttestations: state.numCompleteAttestations + 1,
+        acceptedAttestationCodes: [...state.acceptedAttestationCodes, action.code],
+      }
     case Actions.UPDATE_E164_PHONE_NUMBER_ADDRESSES:
       return {
         ...state,
@@ -144,14 +234,6 @@ export const reducer = (
           ...state.walletToAccountAddress,
           ...action.walletToAccountAddress,
         },
-      }
-    case Actions.REMOVE_WALLET_TO_ACCOUNT_ADDRESS:
-      return {
-        ...state,
-        walletToAccountAddress: removeKeyFromMapping(
-          state.walletToAccountAddress,
-          action.walletAddress
-        ),
       }
     case Actions.UPDATE_E164_PHONE_NUMBER_SALT:
       return {
@@ -274,11 +356,60 @@ export const reducer = (
         matchedContacts: state.matchedContacts,
         secureSendPhoneNumberMapping: state.secureSendPhoneNumberMapping,
       }
+    case Actions.SET_LAST_REVEAL_ATTEMPT:
+      return {
+        ...state,
+        lastRevealAttempt: action.time,
+      }
+    case Actions.SET_ATTESTATION_INPUT_STATUS:
+      return {
+        ...state,
+        attestationInputStatus: updatedInputStatuses(state, action.index, action.status),
+      }
     default:
       return state
   }
 }
 
+function updatedInputStatuses(state: State, index: number, status: CodeInputStatus) {
+  const newStatuses = [...state.attestationInputStatus]
+  newStatuses[index] = status
+  Logger.debug('identityReducer@attestationInputStatus', newStatuses)
+  return newStatuses
+}
+
+const completeCodeReducer = (state: State, numCompleteAttestations: number) => {
+  const { attestationCodes, acceptedAttestationCodes } = state
+  // Ensure numCompleteAttestations many codes are filled
+  const updatedAttestationCodes: AttestationCode[] = []
+  for (let i = 0; i < numCompleteAttestations; i++) {
+    updatedAttestationCodes[i] = acceptedAttestationCodes[i] || {
+      code: ATTESTATION_CODE_PLACEHOLDER,
+      issuer: ATTESTATION_ISSUER_PLACEHOLDER,
+    }
+  }
+  for (let i = numCompleteAttestations; i < NUM_ATTESTATIONS_REQUIRED; i++) {
+    if (
+      attestationCodes[i] &&
+      !isCodeRepeated(
+        attestationCodes.filter((code) => code).map((code) => code.code),
+        attestationCodes[i].code
+      )
+    ) {
+      updatedAttestationCodes.push(attestationCodes[i])
+    }
+  }
+  return {
+    numCompleteAttestations,
+    attestationCodes: updatedAttestationCodes,
+  }
+}
+
+export const attestationCodesSelector = (state: RootState) => state.identity.attestationCodes
+export const acceptedAttestationCodesSelector = (state: RootState) =>
+  state.identity.acceptedAttestationCodes
+export const attestationInputStatusSelector = (state: RootState) =>
+  state.identity.attestationInputStatus
 export const e164NumberToAddressSelector = (state: RootState) => state.identity.e164NumberToAddress
 export const addressToE164NumberSelector = (state: RootState) => state.identity.addressToE164Number
 export const walletToAccountAddressSelector = (state: RootState) =>
