+++ packages/mobile/src/identity/smsRetrieval.ts
@@ -1,8 +1,9 @@
 import SmsRetriever from '@celo/react-native-sms-retriever'
 import { eventChannel } from 'redux-saga'
 import { call, put, take } from 'redux-saga/effects'
+import { receiveAttestationMessage } from 'src/identity/actions'
+import { CodeInputType, NUM_ATTESTATIONS_REQUIRED } from 'src/identity/verification'
 import Logger from 'src/utils/Logger'
-import { CodeInputType, NUM_ATTESTATIONS_REQUIRED, receiveAttestationCode } from 'src/verify/module'
 
 const TAG = 'identity/smsRetrieval'
 
@@ -24,7 +25,7 @@ export function* startAutoSmsRetrieval() {
       const { message } = yield take(autoSmsChannel)
       if (!messages.includes(message)) {
         messages.push(message)
-        yield put(receiveAttestationCode({ message, inputType: CodeInputType.AUTOMATIC }))
+        yield put(receiveAttestationMessage(message, CodeInputType.AUTOMATIC))
       }
     }
   } catch (error) {
