+++ packages/mobile/src/identity/types.ts
@@ -4,6 +4,18 @@ export enum RecipientVerificationStatus {
   UNKNOWN = 2,
 }
 
+export enum VerificationStatus {
+  SaltQuotaExceeded = -3,
+  InsufficientBalance = -2,
+  Failed = -1,
+  Stopped = 0,
+  Prepping = 1,
+  GettingStatus = 2,
+  RequestingAttestations = 3,
+  CompletingAttestations = 4,
+  Done = 6,
+}
+
 export enum ImportContactsStatus {
   Failed = -1,
   Stopped = 0,
