+++ packages/mobile/src/verify/VerificationEducationScreen.test.tsx
@@ -7,7 +7,7 @@ import { ErrorMessages } from 'src/app/ErrorMessages'
 import { features } from 'src/flags'
 import { Screens } from 'src/navigator/Screens'
 import { useAsyncKomenciReadiness } from 'src/verify/hooks'
-import { idle, KomenciAvailable } from 'src/verify/module'
+import { idle, KomenciAvailable } from 'src/verify/reducer'
 import VerificationEducationScreen from 'src/verify/VerificationEducationScreen'
 import { createMockStore, getMockStackScreenProps } from 'test/utils'
