+++ packages/mobile/src/verify/VerificationInputScreen.tsx
@@ -21,9 +21,19 @@ import { errorSelector } from 'src/alert/reducer'
 import { ErrorMessages } from 'src/app/ErrorMessages'
 import { shortVerificationCodesEnabledSelector } from 'src/app/selectors'
 import BackButton from 'src/components/BackButton'
+import { CodeInputStatus } from 'src/components/CodeInput'
 import DevSkipButton from 'src/components/DevSkipButton'
 import { ALERT_BANNER_DURATION, ATTESTATION_REVEAL_TIMEOUT_SECONDS } from 'src/config'
 import i18n, { Namespaces, withTranslation } from 'src/i18n'
+import {
+  cancelVerification,
+  receiveAttestationMessage,
+  resendAttestations,
+  setAttestationInputStatus,
+} from 'src/identity/actions'
+import { attestationInputStatusSelector } from 'src/identity/reducer'
+import { VerificationStatus } from 'src/identity/types'
+import { CodeInputType, NUM_ATTESTATIONS_REQUIRED } from 'src/identity/verification'
 import { HeaderTitleWithSubtitle, nuxNavigationOptions } from 'src/navigator/Headers'
 import { navigate, navigateHome } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
@@ -32,17 +42,6 @@ import TopBarTextButtonOnboarding from 'src/onboarding/TopBarTextButtonOnboardin
 import { RootState } from 'src/redux/reducers'
 import Logger from 'src/utils/Logger'
 import { timeDeltaInSeconds } from 'src/utils/time'
-import {
-  AttestationCode,
-  cancel as cancelVerification,
-  CodeInputType,
-  NUM_ATTESTATIONS_REQUIRED,
-  OnChainVerificationStatus,
-  receiveAttestationCode,
-  resendMessages,
-  setAttestationInputStatus,
-  VerificationState,
-} from 'src/verify/module'
 import { isCodeRepeated } from 'src/verify/utils'
 import VerificationCodeInput from 'src/verify/VerificationCodeInput'
 import VerificationInputHelpDialog from 'src/verify/VerificationInputHelpDialog'
@@ -53,9 +52,9 @@ type ScreenProps = StackScreenProps<StackParamList, Screens.VerificationInputScr
 
 interface StateProps {
   e164Number: string | null
-  attestationCodes: AttestationCode[]
-  verificationStatus: OnChainVerificationStatus
-  verificationState: VerificationState
+  attestationInputStatus: CodeInputStatus[]
+  numCompleteAttestations: number
+  verificationStatus: VerificationStatus
   underlyingError: ErrorMessages | null | undefined
   lastRevealAttempt: number | null
   shortVerificationCodesEnabled: boolean
@@ -63,8 +62,8 @@ interface StateProps {
 
 interface DispatchProps {
   cancelVerification: typeof cancelVerification
-  receiveAttestationCode: typeof receiveAttestationCode
-  resendMessages: typeof resendMessages
+  receiveAttestationMessage: typeof receiveAttestationMessage
+  resendAttestations: typeof resendAttestations
   setAttestationInputStatus: typeof setAttestationInputStatus
   hideAlert: typeof hideAlert
   showMessage: typeof showMessage
@@ -81,8 +80,8 @@ interface State {
 
 const mapDispatchToProps = {
   cancelVerification,
-  receiveAttestationCode,
-  resendMessages,
+  receiveAttestationMessage,
+  resendAttestations,
   setAttestationInputStatus,
   hideAlert,
   showMessage,
@@ -90,14 +89,14 @@ const mapDispatchToProps = {
 }
 
 const mapStateToProps = (state: RootState): StateProps => {
-  const attestationCodes = state.verify.attestationCodes
-  const lastRevealAttempt = state.verify.lastRevealAttempt
+  const numCompleteAttestations = state.identity.numCompleteAttestations
+  const lastRevealAttempt = state.identity.lastRevealAttempt
 
   return {
     e164Number: state.account.e164PhoneNumber,
-    attestationCodes,
-    verificationState: state.verify.currentState,
-    verificationStatus: state.verify.status,
+    attestationInputStatus: attestationInputStatusSelector(state),
+    numCompleteAttestations,
+    verificationStatus: state.identity.verificationStatus,
     underlyingError: errorSelector(state),
     shortVerificationCodesEnabled: shortVerificationCodesEnabledSelector(state),
     lastRevealAttempt,
@@ -168,8 +167,8 @@ class VerificationInputScreen extends React.Component<Props, State> {
 
   isVerificationComplete = (prevProps: Props) => {
     return (
-      prevProps.verificationStatus.completed < NUM_ATTESTATIONS_REQUIRED &&
-      this.props.verificationStatus.completed >= NUM_ATTESTATIONS_REQUIRED
+      prevProps.numCompleteAttestations < NUM_ATTESTATIONS_REQUIRED &&
+      this.props.numCompleteAttestations >= NUM_ATTESTATIONS_REQUIRED
     )
   }
 
@@ -191,12 +190,7 @@ class VerificationInputScreen extends React.Component<Props, State> {
         ((shortVerificationCodesEnabled && extractSecurityCodeWithPrefix(value)) ||
           extractAttestationCodeFromMessage(value))
       ) {
-        this.setState((state) => dotProp.set(state, `codeSubmittingStatuses.${index}`, true))
-        this.props.receiveAttestationCode({
-          message: value,
-          inputType: CodeInputType.MANUAL,
-          index,
-        })
+        this.props.receiveAttestationMessage(value, CodeInputType.MANUAL, index)
       }
     }
   }
@@ -226,7 +220,7 @@ class VerificationInputScreen extends React.Component<Props, State> {
       timeDeltaInSeconds(Date.now(), lastRevealAttempt) > ATTESTATION_REVEAL_TIMEOUT_SECONDS
 
     if (isRevealAllowed) {
-      this.props.resendMessages()
+      this.props.resendAttestations()
     } else {
       this.props.showMessage(
         this.props.t('verificationPrematureRevealMessage'),
@@ -238,7 +232,7 @@ class VerificationInputScreen extends React.Component<Props, State> {
 
   render() {
     const { codeInputValues, isKeyboardVisible, timer } = this.state
-    const { t, route, shortVerificationCodesEnabled, verificationStatus } = this.props
+    const { t, numCompleteAttestations, route, shortVerificationCodesEnabled } = this.props
 
     const showHelpDialog = route.params?.showHelpDialog || false
     const translationPlatformContext = Platform.select({ ios: 'ios' })
@@ -292,10 +286,8 @@ class VerificationInputScreen extends React.Component<Props, State> {
                     <View style={styles.spacer} />
                     <TextButton style={styles.resendButton} onPress={this.onPressResend}>
                       {t(
-                        `verificationInput.resendMessages${
-                          verificationStatus.completed ? '' : '_all'
-                        }`,
-                        { count: NUM_ATTESTATIONS_REQUIRED - verificationStatus.completed }
+                        `verificationInput.resendMessages${numCompleteAttestations ? '' : '_all'}`,
+                        { count: NUM_ATTESTATIONS_REQUIRED - numCompleteAttestations }
                       )}
                     </TextButton>
                   </KeyboardAwareScrollView>
