+++ packages/mobile/src/verify/VerificationLoadingScreen.test.tsx
@@ -1,25 +1,28 @@
 import * as React from 'react'
 import { render } from 'react-native-testing-library'
 import { Provider } from 'react-redux'
-import { error, idle } from 'src/verify/module'
+import { VerificationStatus } from 'src/identity/types'
+import { Screens } from 'src/navigator/Screens'
 import VerificationLoadingScreen from 'src/verify/VerificationLoadingScreen'
 import { createMockStore } from 'test/utils'
+import { mockNavigation } from 'test/values'
 
 // Lock time so snapshots always show the same countdown value
 jest.spyOn(Date, 'now').mockImplementation(() => 1487076708000)
 
+const mockRoute = {
+  name: Screens.VerificationLoadingScreen as Screens.VerificationLoadingScreen,
+  key: '1',
+  params: {
+    withoutRevealing: false,
+  },
+}
+
 describe('VerificationLoadingScreen', () => {
   it('renders correctly', () => {
     const { toJSON } = render(
-      <Provider
-        store={createMockStore({
-          verify: {
-            currentState: idle(),
-            attestationCodes: [],
-          },
-        })}
-      >
-        <VerificationLoadingScreen />
+      <Provider store={createMockStore()}>
+        <VerificationLoadingScreen navigation={mockNavigation} route={mockRoute} />
       </Provider>
     )
     expect(toJSON()).toMatchSnapshot()
@@ -27,14 +30,13 @@ describe('VerificationLoadingScreen', () => {
 
   it('renders correctly with fail modal', () => {
     const store = createMockStore({
-      verify: {
-        currentState: error('test'),
-        attestationCodes: [],
+      identity: {
+        verificationStatus: VerificationStatus.Failed,
       },
     })
     const { toJSON } = render(
       <Provider store={store}>
-        <VerificationLoadingScreen />
+        <VerificationLoadingScreen navigation={mockNavigation} route={mockRoute} />
       </Provider>
     )
     expect(toJSON()).toMatchSnapshot()
