+++ packages/mobile/src/verify/VerificationLoadingScreen.tsx
@@ -4,6 +4,7 @@ import colors from '@celo/react-components/styles/colors'
 import fontStyles from '@celo/react-components/styles/fonts'
 import { Spacing } from '@celo/react-components/styles/styles'
 import { useFocusEffect, useIsFocused } from '@react-navigation/native'
+import { StackScreenProps } from '@react-navigation/stack'
 import LottieView from 'lottie-react-native'
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
@@ -18,6 +19,8 @@ import { shallowEqual, useDispatch, useSelector } from 'react-redux'
 import CancelButton from 'src/components/CancelButton'
 import Carousel, { CarouselItem } from 'src/components/Carousel'
 import { Namespaces } from 'src/i18n'
+import { cancelVerification } from 'src/identity/actions'
+import { VerificationStatus } from 'src/identity/types'
 import {
   verificationEducation1,
   verificationEducation2,
@@ -27,15 +30,11 @@ import {
 import { noHeaderGestureDisabled } from 'src/navigator/Headers'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
+import { StackParamList } from 'src/navigator/types'
 import { RootState } from 'src/redux/reducers'
 import Logger from 'src/utils/Logger'
 import useBackHandler from 'src/utils/useBackHandler'
 import AlternatingText from 'src/verify/AlternatingText'
-import {
-  cancel as cancelVerification,
-  VerificationState,
-  VerificationStateType,
-} from 'src/verify/module'
 import VerificationCountdown from 'src/verify/VerificationCountdown'
 import { VerificationFailedModal } from 'src/verify/VerificationFailedModal'
 
@@ -44,15 +43,17 @@ const TAG = 'VerificationLoadingScreen'
 const mapStateToProps = (state: RootState) => {
   return {
     e164Number: state.account.e164PhoneNumber,
-    verificationState: state.verify.currentState,
+    verificationStatus: state.identity.verificationStatus,
     retryWithForno: state.account.retryVerificationWithForno,
     fornoMode: state.web3.fornoMode,
   }
 }
 
-export default function VerificationLoadingScreen() {
-  const verificationStateRef = useRef<VerificationState | undefined>()
-  const { fornoMode, retryWithForno, verificationState } = useSelector(
+type Props = StackScreenProps<StackParamList, Screens.VerificationLoadingScreen>
+
+export default function VerificationLoadingScreen({ route }: Props) {
+  const verificationStatusRef = useRef<VerificationStatus | undefined>()
+  const { fornoMode, retryWithForno, verificationStatus } = useSelector(
     mapStateToProps,
     shallowEqual
   )
@@ -69,17 +70,17 @@ export default function VerificationLoadingScreen() {
   )
 
   useEffect(() => {
-    if (!isFocused || verificationStateRef.current === verificationState) {
+    if (!isFocused || verificationStatusRef.current === verificationStatus) {
       return
     }
-    verificationStateRef.current = verificationState
+    verificationStatusRef.current = verificationStatus
 
-    if (verificationState.type === VerificationStateType.CompletingAttestations) {
+    if (verificationStatus === VerificationStatus.CompletingAttestations) {
       navigate(Screens.VerificationInputScreen)
-    } else if (verificationState.type === VerificationStateType.Success) {
+    } else if (verificationStatus === VerificationStatus.Done) {
       navigate(Screens.ImportContacts)
     }
-  }, [verificationState, isFocused])
+  }, [verificationStatus, isFocused])
 
   useBackHandler(() => {
     // Cancel verification when user presses back button on this screen
@@ -95,8 +96,8 @@ export default function VerificationLoadingScreen() {
 
   const onFinishCountdown = () => {
     // For now switch to the verification screen
-    // if we haven't reached the completion stage yet
-    if (!isFocused || verificationState.type === VerificationStateType.CompletingAttestations) {
+    // if we haven't reached the reveal stage yet
+    if (!isFocused || verificationStatus === VerificationStatus.CompletingAttestations) {
       return
     }
     navigate(Screens.VerificationInputScreen)
@@ -226,7 +227,9 @@ export default function VerificationLoadingScreen() {
               primaryText={t('loading.confirming')}
               secondaryText={t('loading.pleaseKeepAppOpen')}
             />
-            <VerificationCountdown startTime={countdownStartTime} onFinish={onFinishCountdown} />
+            {!route.params.withoutRevealing && (
+              <VerificationCountdown startTime={countdownStartTime} onFinish={onFinishCountdown} />
+            )}
           </Animated.View>
           <Animated.View style={learnMoreContainerStyle}>
             <TouchableOpacity style={styles.upHandleContainer} onPress={onPressLearnMore}>
@@ -246,7 +249,7 @@ export default function VerificationLoadingScreen() {
         />
       </Animated.ScrollView>
       <VerificationFailedModal
-        verificationState={verificationState}
+        verificationStatus={verificationStatus}
         retryWithForno={retryWithForno}
         fornoMode={fornoMode}
       />
