--- packages/mobile/src/verify/komenci.ts
@@ -1,415 +1,0 @@
-import { Result } from '@celo/base/lib/result'
-import { Address, ContractKit } from '@celo/contractkit/lib'
-import { AttestationsWrapper } from '@celo/contractkit/lib/wrappers/Attestations'
-import { CheckSessionResp, StartSessionResp } from '@celo/komencikit/src/actions'
-import {
-  AuthenticationFailed,
-  FetchError,
-  InvalidWallet,
-  KomenciDown,
-  LoginSignatureError,
-  NetworkError,
-  NotFoundError,
-  RequestError,
-  ResponseDecodeError,
-  ServiceUnavailable,
-  TxError,
-  TxEventNotFound,
-  TxRevertError,
-  TxTimeoutError,
-  WalletValidationError,
-} from '@celo/komencikit/src/errors'
-import { KomenciKit } from '@celo/komencikit/src/kit'
-import { verifyWallet } from '@celo/komencikit/src/verifyWallet'
-import { sleep } from '@celo/utils/lib/async'
-import { AttestationsStatus } from '@celo/utils/lib/attestations'
-import { all, call, put, select } from 'redux-saga/effects'
-import networkConfig from 'src/geth/networkConfig'
-import {
-  hasExceededKomenciErrorQuota,
-  KomenciErrorQuotaExceeded,
-  KomenciSessionInvalidError,
-  storeTimestampIfKomenciError,
-} from 'src/identity/feelessVerificationErrors'
-import { e164NumberToSaltSelector, E164NumberToSaltType } from 'src/identity/reducer'
-
-import { navigate } from 'src/navigator/NavigationService'
-import { Screens } from 'src/navigator/Screens'
-import Logger from 'src/utils/Logger'
-import {
-  e164NumberSelector,
-  fail,
-  fetchOnChainData,
-  fetchPhoneNumberDetails,
-  KomenciAvailable,
-  KomenciContext,
-  komenciContextSelector,
-  phoneHashSelector,
-  setKomenciAvailable,
-  setKomenciContext,
-  setVerificationStatus,
-  shouldUseKomenciSelector,
-  start,
-} from 'src/verify/module'
-import { getAttestationsStatus } from 'src/verify/saga'
-import { getContractKit } from 'src/web3/contracts'
-import { registerWalletAndDekViaKomenci } from 'src/web3/dataEncryptionKey'
-import { getAccount, getConnectedUnlockedAccount } from 'src/web3/saga'
-
-const TAG = 'verify/komenci'
-
-const KOMENCI_READINESS_RETRIES = 3
-const KOMENCI_DEPLOY_MTW_RETRIES = 3
-
-export function* getKomenciAwareAccount() {
-  const komenci: KomenciContext = yield select(komenciContextSelector)
-  const { unverifiedMtwAddress } = komenci
-  const shouldUseKomenci = yield select(shouldUseKomenciSelector)
-  return shouldUseKomenci && unverifiedMtwAddress
-    ? unverifiedMtwAddress
-    : yield call(getConnectedUnlockedAccount)
-}
-
-export function* checkIfKomenciAvailableSaga() {
-  Logger.debug(TAG, '@checkIfKomenciAvailableSaga')
-  const contractKit: ContractKit = yield call(getContractKit)
-  const walletAddress: Address = yield call(getAccount)
-  const komenci: KomenciContext = yield select(komenciContextSelector)
-  const komenciKit: KomenciKit = yield call(getKomenciKit, contractKit, walletAddress, komenci)
-
-  const isKomenciAvailable: boolean = yield call(fetchKomenciReadiness, komenciKit)
-  yield put(setKomenciAvailable(isKomenciAvailable ? KomenciAvailable.Yes : KomenciAvailable.No))
-}
-
-export function* startOrResumeKomenciSessionSaga() {
-  // TODO: Move this out of saga
-  yield call(navigate, Screens.VerificationLoadingScreen)
-
-  Logger.debug(TAG, '@startOrResumeKomenciSession', 'Starting session')
-
-  const contractKit: ContractKit = yield call(getContractKit)
-  const walletAddress: Address = yield call(getConnectedUnlockedAccount)
-  const komenci: KomenciContext = yield select(komenciContextSelector)
-  const komenciKit: KomenciKit = yield call(getKomenciKit, contractKit, walletAddress, komenci)
-
-  const { sessionActive, captchaToken } = komenci
-  let { sessionToken } = komenci
-
-  // If there isn't an active session, start one. Need to include `sessionActive`
-  // because that's the only way we'll know if Komenci session is active but
-  // quota is used
-  if (!sessionActive || !sessionToken.length) {
-    // Should never get here without a captcha token
-    if (!captchaToken.length) {
-      const error = new KomenciSessionInvalidError()
-      Logger.error(TAG, '@startOrResumeKomenciSession', error)
-      throw error
-    }
-
-    const komenciSessionResult: Result<
-      StartSessionResp,
-      FetchError | AuthenticationFailed | LoginSignatureError
-    > = yield call([komenciKit, komenciKit.startSession], captchaToken)
-
-    if (!komenciSessionResult.ok) {
-      Logger.debug(TAG, '@startOrResumeKomenciSession', 'Unable to start session')
-      throw komenciSessionResult.error
-    }
-
-    sessionToken = komenciSessionResult.result.token
-
-    yield put(
-      setKomenciContext({
-        sessionToken: komenciSessionResult.result.token,
-        callbackUrl: komenciSessionResult.result.callbackUrl || '',
-        sessionActive: true,
-      })
-    )
-  }
-
-  Logger.debug(TAG, 'Session active. sessionToken: ', sessionToken)
-  yield put(fetchPhoneNumberDetails())
-}
-
-export function* fetchOrDeployMtwSaga() {
-  const e164Number = yield select(e164NumberSelector)
-  const contractKit = yield call(getContractKit)
-  const walletAddress = yield call(getConnectedUnlockedAccount)
-  const komenci = yield select(komenciContextSelector)
-  const komenciKit = yield call(getKomenciKit, contractKit, walletAddress, komenci)
-
-  try {
-    // Now that we are guarnateed to have the phoneHash, check again to see if the
-    // user already has a verified MTW
-    const verifiedMtwAddress = yield call(fetchVerifiedMtw, contractKit, walletAddress)
-    if (verifiedMtwAddress) {
-      // TODO: Address this
-      // yield put(doVerificationFlow(true))
-      return
-    }
-
-    Logger.debug(TAG, '@fetchOrDeployMtwSaga', 'Starting fetch')
-    const storedUnverifiedMtwAddress = komenci.unverifiedMtwAddress
-    let deployedUnverifiedMtwAddress: string | null = null
-    // If there isn't a MTW stored for this session, ask Komenci to deploy one
-    if (!storedUnverifiedMtwAddress) {
-      for (let i = 0; i < KOMENCI_DEPLOY_MTW_RETRIES; i += 1) {
-        // This try/catch block is a workaround because Komenci will throw an error
-        // if a wallet was already deployed in a session. This is only fatal if
-        // we can't recover the MTW address or there is no quota left on the session
-        try {
-          const deployWalletResult: Result<
-            string,
-            FetchError | TxError | InvalidWallet
-          > = yield call(
-            [komenciKit, komenciKit.deployWallet],
-            networkConfig.currentMtwImplementationAddress
-          )
-
-          if (!deployWalletResult.ok) {
-            Logger.debug(TAG, '@fetchOrDeployMtw', 'Unable to deploy MTW')
-            throw deployWalletResult.error
-          }
-          deployedUnverifiedMtwAddress = deployWalletResult.result
-          break
-        } catch (e) {
-          storeTimestampIfKomenciError(e)
-
-          switch (true) {
-            case e instanceof ServiceUnavailable:
-            case e instanceof NetworkError:
-            case e instanceof RequestError:
-            case e instanceof NotFoundError:
-            case e instanceof ResponseDecodeError:
-            case e instanceof TxTimeoutError:
-            case e instanceof TxRevertError:
-            case e instanceof TxEventNotFound:
-              continue
-
-            case e instanceof InvalidWallet:
-            default:
-              put(setKomenciAvailable(KomenciAvailable.No))
-              put(start({ e164Number }))
-              return
-          }
-        }
-      }
-    }
-
-    const unverifiedMtwAddress = deployedUnverifiedMtwAddress ?? storedUnverifiedMtwAddress
-
-    // If we couldn't recover or deploy a new the MTW address, then propogate the Komenci error
-    // we recevied from the failed `deployWallet` call. We also need to check if the session
-    // is still active because it's possible the current session ran out of quota
-    if (!unverifiedMtwAddress || !komenci.sessionActive) {
-      Logger.debug(TAG, '@fetchOrDeployMtw', 'Unable to deploy or recover a MTW')
-      // The new error on the RHS is mostly to placate the linting rules.
-      // There should be no instances where Komenci is unable to deploy
-      // a MTW yet doesn't return an error
-      throw new Error('Unable to deploy or recover a MTW')
-    }
-
-    // Check if the MTW we have is a valid implementation
-    const validityCheckResult: Result<true, WalletValidationError> = yield call(
-      verifyWallet,
-      contractKit,
-      unverifiedMtwAddress,
-      networkConfig.allowedMtwImplementations,
-      walletAddress
-    )
-
-    if (!validityCheckResult.ok) {
-      Logger.debug(TAG, '@fetchOrDeployMtw', 'Unable to validate MTW implementation')
-      throw validityCheckResult.error
-    }
-
-    yield put(setKomenciContext({ unverifiedMtwAddress }))
-    yield call(feelessDekAndWalletRegistration, komenciKit, walletAddress, unverifiedMtwAddress)
-    yield put(fetchOnChainData())
-  } catch (e) {
-    storeTimestampIfKomenciError(e)
-    put(setKomenciAvailable(KomenciAvailable.No))
-    put(start(e164Number))
-  }
-}
-
-export function* fetchKomenciReadiness(komenciKit: KomenciKit) {
-  for (let i = 0; i < KOMENCI_READINESS_RETRIES; i += 1) {
-    const serviceStatusResult: Result<true, KomenciDown> = yield call([
-      komenciKit,
-      komenciKit.checkService,
-    ])
-
-    if (!serviceStatusResult.ok) {
-      Logger.debug(TAG, '@fetchKomenciReadiness', 'Komenci service is down')
-      yield call(checkTooManyErrors)
-      if (serviceStatusResult.error instanceof KomenciDown) {
-        yield sleep(2 ** i * 5000)
-      } else {
-        throw serviceStatusResult.error
-      }
-    }
-  }
-
-  return true
-}
-
-export function* fetchKomenciSession(komenciKit: KomenciKit, e164Number: string) {
-  Logger.debug(TAG, '@fetchKomenciSession', 'Starting fetch')
-  const [komenciContext, pepperCache, sessionStatusResult]: [
-    KomenciContext,
-    E164NumberToSaltType,
-    Result<CheckSessionResp, FetchError>
-  ] = yield all([
-    select(komenciContextSelector),
-    select(e164NumberToSaltSelector),
-    call([komenciKit, komenciKit.checkSession]),
-  ])
-
-  let sessionActive = true
-  let { unverifiedMtwAddress } = komenciContext
-
-  // An inactive session is not fatal, it just means we will need to start one
-  if (!sessionStatusResult.ok) {
-    Logger.debug(TAG, '@fetchKomenciSession', 'No active sessions')
-    sessionActive = false
-  } else {
-    Logger.debug(TAG, '@fetchKomenciSession', 'Active session found')
-    const {
-      quotaLeft: { distributedBlindedPepper, requestSubsidisedAttestation, submitMetaTransaction },
-      metaTxWalletAddress,
-    } = sessionStatusResult.result
-    const ownPepper = pepperCache[e164Number]
-
-    Logger.debug(
-      TAG,
-      '@fetchKomenciSession Session status:',
-      JSON.stringify(sessionStatusResult.result)
-    )
-
-    // Sometimes `metaTxWalletAddress` is returned as undefined for an active session.
-    // In that case, use the `unverifiedMtwAddress` we have stored locally
-    unverifiedMtwAddress = metaTxWalletAddress ?? unverifiedMtwAddress
-
-    // No pepper quota remaining is only bad if it's not already cached. Given Komenci will fetch
-    // a pepper for you once, a session could be invalid due to the pepper condition if a user
-    // fetched their pepper once this session then uninstalled without starting a new session
-    if (
-      (!ownPepper && !distributedBlindedPepper) ||
-      !requestSubsidisedAttestation ||
-      !submitMetaTransaction
-    ) {
-      Logger.debug(
-        TAG,
-        '@fetchKomenciSession',
-        'Komenci session has run out of quota. Will attempt to start a new one'
-      )
-      sessionActive = false
-    }
-  }
-
-  yield put(setKomenciContext({ unverifiedMtwAddress, sessionActive }))
-}
-
-export function getKomenciKit(
-  contractKit: ContractKit,
-  walletAddress: Address,
-  komenci: KomenciContext
-) {
-  return new KomenciKit(contractKit, walletAddress, {
-    url: komenci.callbackUrl || networkConfig.komenciUrl,
-    token: komenci.sessionToken,
-  })
-}
-
-function* checkTooManyErrors() {
-  const komenci: KomenciContext = yield select(komenciContextSelector)
-  if (hasExceededKomenciErrorQuota(komenci.errorTimestamps)) {
-    Logger.debug(TAG, '@fetchKomenciReadiness', 'Too  many errors')
-    throw new KomenciErrorQuotaExceeded()
-  }
-}
-
-function* fetchVerifiedMtw(contractKit: ContractKit, walletAddress: string) {
-  Logger.debug(TAG, '@fetchVerifiedMtw', 'Starting fetch')
-  const phoneHash = yield select(phoneHashSelector)
-
-  const attestationsWrapper: AttestationsWrapper = yield call([
-    contractKit.contracts,
-    contractKit.contracts.getAttestations,
-  ])
-
-  const associatedAccounts: string[] = yield call(
-    [attestationsWrapper, attestationsWrapper.lookupAccountsForIdentifier],
-    phoneHash
-  )
-
-  const accountAttestationStatuses: AttestationsStatus[] = yield all(
-    associatedAccounts.map((account) =>
-      call(getAttestationsStatus, attestationsWrapper, account, phoneHash)
-    )
-  )
-
-  const possibleMtwAddressIndexes: number[] = associatedAccounts
-    .map((_, i) => i)
-    .filter((i) => accountAttestationStatuses[i].isVerified)
-
-  if (!possibleMtwAddressIndexes.length) {
-    Logger.debug(TAG, '@fetchVerifiedMtw', 'No possible MTWs found')
-    return null
-  }
-
-  const verificationResults: Array<Result<true, WalletValidationError>> = yield all(
-    possibleMtwAddressIndexes.map((possibleMtwAddressIndex) =>
-      call(
-        verifyWallet,
-        contractKit,
-        associatedAccounts[possibleMtwAddressIndex],
-        networkConfig.allowedMtwImplementations,
-        walletAddress
-      )
-    )
-  )
-
-  const verifiedMtwAddressIndexes = possibleMtwAddressIndexes.filter(
-    (_, i) => verificationResults[i].ok
-  )
-
-  if (verifiedMtwAddressIndexes.length > 1) {
-    yield put(
-      fail('More than one verified MTW with walletAddress as signer found. Should never happen')
-    )
-    return
-  }
-  if (!verifiedMtwAddressIndexes.length) {
-    Logger.debug(TAG, '@fetchVerifiedMtw', 'No verified MTW found')
-    return null
-  }
-
-  const verifiedMtwAddress = associatedAccounts[verifiedMtwAddressIndexes[0]]
-  const verifiedMtwStatus = accountAttestationStatuses[verifiedMtwAddressIndexes[0]]
-
-  yield put(
-    setKomenciContext({
-      unverifiedMtwAddress: verifiedMtwAddress,
-    })
-  )
-  yield put(setVerificationStatus(verifiedMtwStatus))
-
-  return verifiedMtwAddress
-}
-
-function* feelessDekAndWalletRegistration(
-  komenciKit: KomenciKit,
-  walletAddress: string,
-  unverifiedMtwAddress: string
-) {
-  Logger.debug(TAG, '@feelessDekAndWalletRegistration', 'Starting registration')
-
-  // Should never happen
-  if (!unverifiedMtwAddress) {
-    throw Error('Tried registering DEK and walletAddress without a MTW')
-  }
-
-  yield call(registerWalletAndDekViaKomenci, komenciKit, unverifiedMtwAddress, walletAddress)
-}
