+++ packages/mobile/src/verify/reducer.ts
@@ -1,33 +1,25 @@
-import { Address } from '@celo/base'
 import { ActionableAttestation } from '@celo/contractkit/lib/wrappers/Attestations'
-import { isBalanceSufficientForSigRetrieval } from '@celo/identity/lib/odis/phone-number-identifier'
 import { AttestationsStatus } from '@celo/utils/lib/attestations'
 import { createAction, createReducer, createSelector } from '@reduxjs/toolkit'
+import { RootState } from 'src/redux/reducers'
+
+import { isBalanceSufficientForSigRetrieval } from '@celo/identity/lib/odis/phone-number-identifier'
 import BigNumber from 'bignumber.js'
-import _ from 'lodash'
-import { CodeInputStatus } from 'src/components/CodeInput'
 import { celoTokenBalanceSelector } from 'src/goldToken/selectors'
 import { getRehydratePayload, REHYDRATE, RehydrateAction } from 'src/redux/persist-helper'
-import { RootState } from 'src/redux/reducers'
 import { stableTokenBalanceSelector } from 'src/stableToken/reducer'
-
-export const NUM_ATTESTATIONS_REQUIRED = 3
-export const REVEAL_RETRY_DELAY = 10 * 1000 // 10 seconds
-export const BALANCE_CHECK_TIMEOUT = 5 * 1000 // 5 seconds
-export const VERIFICATION_TIMEOUT = 10 * 60 * 1000 // 10 minutes
-
-export const ATTESTATION_CODE_PLACEHOLDER = 'ATTESTATION_CODE_PLACEHOLDER'
-export const ATTESTATION_ISSUER_PLACEHOLDER = 'ATTESTATION_ISSUER_PLACEHOLDER'
 
 const ESTIMATED_COST_PER_ATTESTATION = 0.051
 
 const rehydrate = createAction<any>(REHYDRATE)
 
-export const setSeenVerificationNux = createAction<boolean>('VERIFY/SET_SEEN_VERIFICATION_NUX')
 export const setKomenciContext = createAction<Partial<KomenciContext>>('VERIFY/SET_KOMENCI_CONTEXT')
+export const setOverrideWithoutVerification = createAction<boolean | undefined>(
+  'VERIFY/SET_OVERRIDE_WITHOUT_VERIFICATION'
+)
 export const checkIfKomenciAvailable = createAction('VERIFY/CHECK_IF_KOMENCI_AVAILABLE')
 export const setKomenciAvailable = createAction<KomenciAvailable>('VERIFY/SET_KOMENCI_AVAILABLE')
-export const start = createAction<{ e164Number: string }>('VERIFY/START')
+export const start = createAction<{ e164Number: string; withoutRevealing: boolean }>('VERIFY/START')
 export const stop = createAction('VERIFY/STOP')
 export const setUseKomenci = createAction<boolean>('VERIFY/SET_USE_KOMENCI')
 export const ensureRealHumanUser = createAction('VERIFY/ENSURE_REAL_HUMAN_USER')
@@ -35,36 +27,10 @@ export const startKomenciSession = createAction('VERIFY/START_KOMENCI_SESSION')
 export const fetchPhoneNumberDetails = createAction('VERIFY/FETCH_PHONE_NUMBER')
 export const fetchMtw = createAction('VERIFY/FETCH_MTW')
 export const fetchOnChainData = createAction('VERIFY/FETCH_ON_CHAIN_DATA')
-export const requestAttestations = createAction('VERIFY/REQUEST_ATTESTATIONS')
-export const revealAttestations = createAction('VERIFY/REVEAL_ATTESTATIONS')
-export const reportRevealStatus = createAction<{
-  attestationServiceUrl: string
-  account: string
-  issuer: string
-  e164Number: string
-  pepper: string
-}>('VERIFY/REPORT_REVEAL_STATUS')
-export const setAttestationInputStatus = createAction<{ index: number; status: CodeInputStatus }>(
-  'VERIFY/SET_ATTESTATION_INPUT_STATUS'
-)
-export const completeAttestations = createAction('VERIFY/COMPLETE_ATTESTATIONS')
 export const fail = createAction<string>('VERIFY/FAIL')
 export const succeed = createAction('VERIFY/SUCCEED')
+export const doVerificationFlow = createAction<boolean>('VERIFY/DO_VERIFICATION_FLOW')
 export const reset = createAction<{ komenci: boolean }>('VERIFY/RESET')
-export const revoke = createAction('VERIFY/REVOKE')
-export const cancel = createAction('VERIFY/CANCEL')
-export const resendMessages = createAction('VERIFY/RESEND_MESSAGES')
-export const receiveAttestationCode = createAction<{
-  message: string
-  inputType: CodeInputType
-  index?: number
-}>('VERIFY/RECEIVE_ATTESTATION_CODE')
-export const inputAttestationCode = createAction<AttestationCode>('VERIFY/INPUT_ATTESTATION_CODE')
-export const completeAttestationCode = createAction<AttestationCode>(
-  'VERIFY/COMPLETE_ATTESTATION_CODE'
-)
-export const setCompletedCodes = createAction<number>('VERIFY/SET_COMPLETED_CODES')
-
 export const setPhoneHash = createAction<string>('VERIFY/SET_PHONE_HASH')
 export const setVerificationStatus = createAction<Partial<AttestationsStatus>>(
   'VERIFY/SET_VERIFICATION_STATUS'
@@ -72,13 +38,8 @@ export const setVerificationStatus = createAction<Partial<AttestationsStatus>>(
 export const setActionableAttestation = createAction<ActionableAttestation[]>(
   'VERIFY/SET_ACTIONABLE_ATTESTATIONS'
 )
-export const setRevealStatuses = createAction<Record<Address, RevealStatus>>(
-  'VERIFY/SET_REVEAL_STATUSES'
-)
-export const setAllRevealStatuses = createAction<RevealStatus>('VERIFY/SET_ALL_REVEAL_STATUSES')
-export const setLastRevealAttempt = createAction<number>('VERIFY/SET_LAST_REVEAL_ATTEMPT')
 
-export enum VerificationStateType {
+export enum StateType {
   Idle = 'Idle',
   Preparing = 'Preparing',
   EnsuringRealHumanUser = 'EnsuringRealHumanUser',
@@ -86,123 +47,75 @@ export enum VerificationStateType {
   FetchingPhoneNumberDetails = 'FetchingPhoneNumberDetails',
   FetchingMtw = 'FetchingMtw',
   FetchingOnChainData = 'FetchingOnChainData',
-  RequestingAttestations = 'RequestingAttestations',
-  RevealingAttestations = 'RevealingAttestations',
-  CompletingAttestations = 'CompletingAttestations',
   Error = 'Error',
-  Success = 'Success',
-}
-
-export interface AttestationCode {
-  code: string
-  shortCode?: string | null
-  issuer: string
-}
-
-export enum CodeInputType {
-  AUTOMATIC = 'automatic',
-  MANUAL = 'manual',
-  DEEP_LINK = 'deepLink',
 }
 
 // Idle State
 interface Idle {
-  type: VerificationStateType.Idle
+  type: StateType.Idle
 }
-export const idle = (): Idle => ({ type: VerificationStateType.Idle })
+export const idle = (): Idle => ({ type: StateType.Idle })
 
 // PreparingKomenci State
 interface Preparing {
-  type: VerificationStateType.Preparing
+  type: StateType.Preparing
 }
 // {}: Omit<PreparingKomenci, 'type'>
 export const preparing = (): Preparing => ({
-  type: VerificationStateType.Preparing,
+  type: StateType.Preparing,
 })
 
 // EnsuringRealHumanUser State
 interface EnsuringRealHumanUser {
-  type: VerificationStateType.EnsuringRealHumanUser
+  type: StateType.EnsuringRealHumanUser
 }
 export const ensuringRealHumanUser = (): EnsuringRealHumanUser => ({
-  type: VerificationStateType.EnsuringRealHumanUser,
+  type: StateType.EnsuringRealHumanUser,
 })
 
 // StartingKomenciSession State
 interface StartingKomenciSession {
-  type: VerificationStateType.StartingKomenciSession
+  type: StateType.StartingKomenciSession
 }
 export const startingKomenciSession = (): StartingKomenciSession => ({
-  type: VerificationStateType.StartingKomenciSession,
+  type: StateType.StartingKomenciSession,
 })
 
 // FetchingPhoneNumberDetails State
 interface FetchingPhoneNumberDetails {
-  type: VerificationStateType.FetchingPhoneNumberDetails
+  type: StateType.FetchingPhoneNumberDetails
 }
 export const fetchingPhoneNumberDetails = (): FetchingPhoneNumberDetails => ({
-  type: VerificationStateType.FetchingPhoneNumberDetails,
+  type: StateType.FetchingPhoneNumberDetails,
 })
 
 // FetchingMtw State
 interface FetchingMtw {
-  type: VerificationStateType.FetchingMtw
+  type: StateType.FetchingMtw
 }
 export const fetchingMtw = (): FetchingMtw => ({
-  type: VerificationStateType.FetchingMtw,
+  type: StateType.FetchingMtw,
 })
 
 // FetchingVerificationOnChain State
 interface FetchingOnChainData {
-  type: VerificationStateType.FetchingOnChainData
+  type: StateType.FetchingOnChainData
 }
 export const fetchingOnChainData = (): FetchingOnChainData => ({
-  type: VerificationStateType.FetchingOnChainData,
-})
-
-// RequestingAttestations State
-interface RequestingAttestations {
-  type: VerificationStateType.RequestingAttestations
-}
-export const requestingAttestations = (): RequestingAttestations => ({
-  type: VerificationStateType.RequestingAttestations,
-})
-
-// RevealingAttestations State
-interface RevealingAttestations {
-  type: VerificationStateType.RevealingAttestations
-}
-export const revealingAttestations = (): RevealingAttestations => ({
-  type: VerificationStateType.RevealingAttestations,
-})
-
-// CompletingAttestations State
-interface CompletingAttestations {
-  type: VerificationStateType.CompletingAttestations
-}
-export const completingAttestations = (): CompletingAttestations => ({
-  type: VerificationStateType.CompletingAttestations,
+  type: StateType.FetchingOnChainData,
 })
 
 // Error State
 interface Error {
-  type: VerificationStateType.Error
+  type: StateType.Error
   message: string
 }
 export const error = (message: string): Error => ({
-  type: VerificationStateType.Error,
+  type: StateType.Error,
   message,
 })
 
-// Succees State
-interface Success {
-  type: VerificationStateType.Success
-}
-export const success = (): Success => ({
-  type: VerificationStateType.Success,
-})
-
-export type VerificationState =
+type InternalState =
   | Idle
   | Preparing
   | EnsuringRealHumanUser
@@ -210,11 +123,7 @@ export type VerificationState =
   | FetchingPhoneNumberDetails
   | FetchingMtw
   | FetchingOnChainData
-  | RequestingAttestations
-  | RevealingAttestations
-  | CompletingAttestations
   | Error
-  | Success
 
 export interface KomenciContext {
   errorTimestamps: number[]
@@ -231,37 +140,20 @@ export enum KomenciAvailable {
   Unknown = 'UNKNOWN',
 }
 
-export enum RevealStatus {
-  NotRevealed = 'NOT_REVEALED',
-  Revealed = 'REVEALED',
-  Failed = 'FAILED',
-}
-
-export type OnChainVerificationStatus = AttestationsStatus & { komenci: boolean }
-
-export type RevealStatuses = Record<Address, RevealStatus>
-
 export interface State {
-  seenVerificationNux: boolean
-  status: OnChainVerificationStatus
+  status: AttestationsStatus & { komenci: boolean }
   actionableAttestations: ActionableAttestation[]
-  revealStatuses: RevealStatuses
-  currentState: VerificationState
+  currentState: InternalState
   komenci: KomenciContext
   komenciAvailable: KomenciAvailable
   phoneHash?: string
   e164Number?: string
-  attestationCodes: AttestationCode[]
-  lastRevealAttempt: number | null
-  // we store acceptedAttestationCodes to tell user if code
-  // was already used even after Actions.RESET_VERIFICATION
-  acceptedAttestationCodes: AttestationCode[]
-  // Represents the status in the UI. Should be of size 3.
-  attestationInputStatus: CodeInputStatus[]
+  retries: number
+  withoutRevealing: boolean
+  TEMPORARY_override_withoutVerification?: boolean
 }
 
 const initialState: State = {
-  seenVerificationNux: false,
   komenci: {
     errorTimestamps: [],
     unverifiedMtwAddress: null,
@@ -278,17 +170,11 @@ const initialState: State = {
     komenci: true,
   },
   actionableAttestations: [],
-  revealStatuses: {},
+  retries: 0,
   currentState: idle(),
   komenciAvailable: KomenciAvailable.Unknown,
-  attestationCodes: [],
-  lastRevealAttempt: null,
-  acceptedAttestationCodes: [],
-  attestationInputStatus: [
-    CodeInputStatus.Inputting,
-    CodeInputStatus.Disabled,
-    CodeInputStatus.Disabled,
-  ],
+  withoutRevealing: false,
+  TEMPORARY_override_withoutVerification: undefined,
 }
 
 export const reducer = createReducer(initialState, (builder) => {
@@ -306,7 +192,6 @@ export const reducer = createReducer(initialState, (builder) => {
         },
         retries: 0,
         currentState: idle(),
-        attestationInputStatus: initialState.attestationInputStatus,
       }
     })
     .addCase(stop, (state) => {
@@ -317,6 +202,7 @@ export const reducer = createReducer(initialState, (builder) => {
         ...state,
         e164Number: action.payload.e164Number,
         currentState: preparing(),
+        withoutRevealing: action.payload.withoutRevealing,
       }
     })
     .addCase(ensureRealHumanUser, (state) => {
@@ -381,23 +267,15 @@ export const reducer = createReducer(initialState, (builder) => {
       }
     })
     .addCase(setActionableAttestation, (state, action) => {
-      const actionableIssuers = action.payload.map((a) => a.issuer)
       return {
         ...state,
-        revealStatuses: _.pick(state.revealStatuses, actionableIssuers),
         actionableAttestations: action.payload,
       }
     })
-    .addCase(setRevealStatuses, (state, action) => {
+    .addCase(setOverrideWithoutVerification, (state, action) => {
       return {
         ...state,
-        revealStatuses: { ...state.revealStatuses, ...action.payload },
-      }
-    })
-    .addCase(setAllRevealStatuses, (state, action) => {
-      return {
-        ...state,
-        revealStatuses: _.mapValues(state.revealStatuses, () => action.payload),
+        TEMPORARY_override_withoutVerification: action.payload,
       }
     })
     .addCase(checkIfKomenciAvailable, (state) => {
@@ -421,109 +299,9 @@ export const reducer = createReducer(initialState, (builder) => {
           komenci: action.payload.komenci,
         },
         komenciAvailable: action.payload.komenci ? KomenciAvailable.Yes : KomenciAvailable.No,
-        attestationCodes: [],
-        numCompleteAttestations: 0,
-      }
-    })
-    .addCase(fail, (state, action) => {
-      return {
-        ...state,
-        currentState: error(action.payload),
-      }
-    })
-    .addCase(succeed, (state) => {
-      return {
-        ...state,
-        currentState: success(),
-      }
-    })
-    .addCase(revoke, () => {
-      return {
-        ...initialState,
-        attestationCodes: [],
-        acceptedAttestationCodes: [],
-        numCompleteAttestations: 0,
-        lastRevealAttempt: null,
-      }
-    })
-    .addCase(setSeenVerificationNux, (state, action) => {
-      return {
-        ...state,
-        seenVerificationNux: action.payload,
-      }
-    })
-    .addCase(setCompletedCodes, (state, action) => {
-      // Ensure action.payload many codes are filled
-      const attestationCodes = []
-      for (let i = 0; i < action.payload; i++) {
-        attestationCodes[i] = state.acceptedAttestationCodes[i] || {
-          code: ATTESTATION_CODE_PLACEHOLDER,
-          issuer: ATTESTATION_ISSUER_PLACEHOLDER,
-        }
-      }
-      return {
-        ...state,
-        attestationCodes,
-      }
-    })
-    .addCase(inputAttestationCode, (state, action) => {
-      return {
-        ...state,
-        attestationCodes: [...state.attestationCodes, action.payload],
-      }
-    })
-    .addCase(completeAttestationCode, (state, action) => {
-      return {
-        ...state,
-        status: {
-          ...state.status,
-          numAttestationsRemaining: state.status.numAttestationsRemaining - 1,
-          completed: state.status.completed + 1,
-        },
-        acceptedAttestationCodes: [...state.acceptedAttestationCodes, action.payload],
-      }
-    })
-    .addCase(requestAttestations, (state) => {
-      return {
-        ...state,
-        currentState: requestingAttestations(),
-      }
-    })
-    .addCase(revealAttestations, (state) => {
-      return {
-        ...state,
-        currentState: revealingAttestations(),
-      }
-    })
-    .addCase(completeAttestations, (state) => {
-      return {
-        ...state,
-        currentState: completingAttestations(),
-      }
-    })
-    .addCase(setLastRevealAttempt, (state, action) => {
-      return {
-        ...state,
-        lastRevealAttempt: action.payload,
-      }
-    })
-    .addCase(setAttestationInputStatus, (state, action) => {
-      return {
-        ...state,
-        attestationInputStatus: updatedInputStatuses(
-          state,
-          action.payload.index,
-          action.payload.status
-        ),
       }
     })
 })
-
-function updatedInputStatuses(state: State, index: number, status: CodeInputStatus) {
-  const newStatuses = [...state.attestationInputStatus]
-  newStatuses[index] = status
-  return newStatuses
-}
 
 const isBalanceSufficientForAttestations = (
   userBalance: BigNumber.Value,
@@ -534,8 +312,6 @@ const isBalanceSufficientForAttestations = (
   )
 }
 
-export const attestationInputStatusSelector = (state: RootState) =>
-  state.verify.attestationInputStatus
 export const currentStateSelector = (state: RootState) => state.verify.currentState
 export const e164NumberSelector = (state: RootState) => state.verify.e164Number
 export const phoneHashSelector = (state: RootState) => state.verify.phoneHash
@@ -553,14 +329,10 @@ export const shouldUseKomenciSelector = (state: RootState) => {
 }
 
 export const verificationStatusSelector = (state: RootState) => state.verify.status
-export const attestationCodesSelector = (state: RootState) => state.verify.attestationCodes
-export const acceptedAttestationCodesSelector = (state: RootState) =>
-  state.verify.acceptedAttestationCodes
 export const actionableAttestationsSelector = (state: RootState): ActionableAttestation[] =>
   state.verify.actionableAttestations
-
-export const revealStatusesSelector = (state: RootState): RevealStatuses =>
-  state.verify.revealStatuses
+export const overrideWithoutVerificationSelector = (state: RootState): boolean | undefined =>
+  state.verify.TEMPORARY_override_withoutVerification
 
 export const isBalanceSufficientForSigRetrievalSelector = createSelector(
   [stableTokenBalanceSelector, celoTokenBalanceSelector],
@@ -591,3 +363,4 @@ export const isBalanceSufficientSelector = createSelector(
     return isBalanceSufficient
   }
 )
+export const withoutRevealingSelector = (state: RootState) => state.verify.withoutRevealing
