+++ packages/mobile/test/schemas.ts
@@ -3,8 +3,9 @@ import { PincodeType } from 'src/account/reducer'
 import { AppState } from 'src/app/actions'
 import { CodeInputStatus } from 'src/components/CodeInput'
 import { DEFAULT_DAILY_PAYMENT_LIMIT_CUSD } from 'src/config'
+import { NUM_ATTESTATIONS_REQUIRED } from 'src/identity/verification'
 import { RootState } from 'src/redux/reducers'
-import { idle, KomenciAvailable, NUM_ATTESTATIONS_REQUIRED } from 'src/verify/module'
+import { idle, KomenciAvailable } from 'src/verify/reducer'
 
 // Default (version -1 schema)
 export const vNeg1Schema = {
@@ -562,24 +563,6 @@ export const v9Schema = {
     walletConnectEnabled: false,
     ranVerificationMigrationAt: null,
   },
-  verify: {
-    ..._.omit(
-      v8Schema.verify,
-      'TEMPORARY_override_withoutVerification',
-      'withoutRevealing',
-      'retries'
-    ),
-    seenVerificationNux: false,
-    revealStatuses: {},
-    attestationCodes: [],
-    acceptedAttestationCodes: [],
-    lastRevealAttempt: null,
-    attestationInputStatus: [
-      CodeInputStatus.Inputting,
-      CodeInputStatus.Disabled,
-      CodeInputStatus.Disabled,
-    ],
-  },
   walletConnect: {
     pairings: [],
     sessions: [],
@@ -589,6 +572,14 @@ export const v9Schema = {
   fiatExchanges: {
     ...v8Schema.fiatExchanges,
     providerLogos: {},
+  },
+  identity: {
+    ...v8Schema.identity,
+    attestationInputStatus: [
+      CodeInputStatus.Inputting,
+      CodeInputStatus.Disabled,
+      CodeInputStatus.Disabled,
+    ],
   },
 }
