+++ packages/mobile/test/values.ts
@@ -1,5 +1,4 @@
 /* Shared mock values to facilitate testing */
-import { ActionableAttestation } from '@celo/contractkit/lib/wrappers/Attestations'
 import { UnlockableWallet } from '@celo/wallet-base'
 import { StackNavigationProp } from '@react-navigation/stack'
 import BigNumber from 'bignumber.js'
@@ -8,6 +7,7 @@ import { TokenTransactionType } from 'src/apollo/types'
 import { EscrowedPayment } from 'src/escrow/actions'
 import { SHORT_CURRENCIES } from 'src/geth/consts'
 import { AddressToE164NumberType, E164NumberToAddressType } from 'src/identity/reducer'
+import { AttestationCode } from 'src/identity/verification'
 import { LocalCurrencyCode } from 'src/localCurrency/consts'
 import { StackParamList } from 'src/navigator/types'
 import { NotificationTypes } from 'src/notifications/types'
@@ -21,24 +21,13 @@ import {
   NumberToRecipient,
   RecipientInfo,
 } from 'src/recipients/recipient'
-import { AttestationCode } from 'src/verify/module'
 
 export const nullAddress = '0x0'
 
 export const mockName = 'John Doe'
 export const mockAccount = '0x0000000000000000000000000000000000007E57'
-export const mockAccount1 = '0x0000000000000000000000000000000000007E58'
 export const mockAccount2 = '0x1Ff482D42D8727258A1686102Fa4ba925C46Bc42'
 export const mockAccount3 = '0x1230000000000000000000000000000000007E57'
-
-export const mockKomenciContext = {
-  errorTimestamps: [],
-  unverifiedMtwAddress: null,
-  sessionActive: false,
-  sessionToken: '',
-  callbackUrl: undefined,
-  captchaToken: '',
-}
 
 export const mockMnemonic =
   'prosper winner find donate tape history measure umbrella agent patrol want rhythm old unable wash wrong need fluid hammer coach reveal plastic trust lake'
@@ -63,12 +52,6 @@ export const mockE164NumberHash =
 export const mockE164NumberPepper = 'piWqRHHYWtfg9'
 export const mockE164NumberHashWithPepper =
   '0xf6429456331dedf8bd32b5e3a578e5bc589a28d012724dcd3e0a4b1be67bb454'
-
-export const mockPhoneHashDetails = {
-  e164Number: mockE164Number,
-  phoneHash: mockE164NumberHash,
-  pepper: mockE164NumberPepper,
-}
 
 export const mockE164Number2 = '+12095559790'
 export const mockDisplayNumber2 = '(209) 555-9790'
@@ -385,47 +368,6 @@ export const mockQRCodeRecipient: AddressRecipient = {
   contactId: undefined,
 }
 
-export const attestationCode0: AttestationCode = {
-  code:
-    'ab8049b95ac02e989aae8b61fddc10fe9b3ac3c6aebcd3e68be495570b2d3da15aabc691ab88de69648f988fab653ac943f67404e532cfd1013627f56365f36501',
-  issuer: '848920b14154b6508b8d98e7ee8159aa84b579a4',
-}
-
-export const attestationCode1: AttestationCode = {
-  code:
-    '2033a9e1268576bf5dfee354a37480529d71f99be82c05005ffb71c7d742d10e7a9aa01f8acc4d7998e1e8b183cf6b8cb4d4a8d923fecfddd191e61e074adc5e00',
-  issuer: 'fdb8da92c3597e81c2737e8be793bee9f1172045',
-}
-
-export const attestationCode2: AttestationCode = {
-  code:
-    '1930a9e1268576bf5dfee354a37480529d71f99be82c05005ffb71c7d742d10e7a9aa01f8acc4d7993f75ab183cf6b8cb4d4a8d923fecfddd191e61e074adc5a10',
-  issuer: 'ecb8da92c3597e81c2737e8be793bee9f1173156',
-}
-
-export const mockActionableAttestations: ActionableAttestation[] = [
-  {
-    issuer: attestationCode0.issuer,
-    blockNumber: 100,
-    attestationServiceURL: 'https://fake.celo.org/0',
-    name: '',
-    version: '1.1.0',
-  },
-  {
-    issuer: attestationCode1.issuer,
-    blockNumber: 110,
-    attestationServiceURL: 'https://fake.celo.org/1',
-    name: '',
-    version: '1.1.0',
-  },
-  {
-    issuer: attestationCode2.issuer,
-    blockNumber: 120,
-    attestationServiceURL: 'https://fake.celo.org/2',
-    name: '',
-    version: '1.1.0',
-  },
-]
 export const mockRecipientInfo: RecipientInfo = {
   phoneRecipientCache: mockPhoneRecipientCache,
   valoraRecipientCache: mockValoraRecipientCache,
